requirements.txt
antlr4-python3-runtime==4.13.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pillow==10.2.0
requests==2.31.0
setuptools==65.5.0
tk==0.1.0
urllib3==2.1.0
